refactor: move types into interface file
